[TASK] Rearrange LogRecordMessages
