+++ Classes/Log/Message/ExceptionHandlerLogRecordMessage.php
@@ -21,32 +21,29 @@ declare(strict_types=1);
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-namespace mteu\StreamWriter\ValueObjects;
+namespace mteu\StreamWriter\Log\Message;
 
 use TYPO3\CMS\Core\Log\LogRecord;
 
 /**
- * LogMessage.
+ * LogOutput.
  *
  * @author Martin Adler <mteu@mailbox.org>
  * @license GPL-3.0-or-later
  */
-final class LogRecordMessage implements Message
+final class ExceptionHandlerLogRecordMessage implements Message
 {
     public function __construct(
         private readonly string $level,
         private readonly string $component,
+        private readonly string $mode,
+        private readonly string $applicationMode,
+        private readonly string $exceptionClass,
+        private readonly int|string $exceptionCode,
+        private readonly string $file,
+        private readonly int|string $line,
         private readonly string $message,
     ) {}
-
-    public static function create(LogRecord $record): self
-    {
-        return new self(
-            $record->getLevel(),
-            $record->getComponent(),
-            $record->getMessage(),
-        );
-    }
 
     public function print(): string
     {
@@ -54,7 +51,37 @@ final class LogRecordMessage implements Message
             '[%s] %s: %s' . PHP_EOL,
             strtoupper($this->level),
             $this->component,
+            $this->printData(),
+        );
+    }
+    public function printData(): string
+    {
+        return sprintf(
+            '(%s: %s) %s, code %d, file %s, line %d: %s',
+            $this->mode,
+            $this->applicationMode,
+            $this->exceptionClass,
+            $this->exceptionCode,
+            $this->file,
+            $this->line,
             $this->message,
+        );
+    }
+
+    public static function create(LogRecord $record): self
+    {
+        $recordData = $record->getData();
+
+        return new self(
+            $record->getLevel(),
+            $record->getComponent(),
+            $recordData['mode'] ?? '',
+            $recordData['application_mode'] ?? '',
+            $recordData['exception_class'] ?? '',
+            $recordData['exception_code'] ?? 0,
+            $recordData['file'] ?? '',
+            $recordData['line'] ?? 0,
+            $recordData['message'] ?? '',
         );
     }
 }
