+++ Classes/Writer/StreamWriter.php
@@ -27,8 +27,8 @@ use mteu\StreamWriter\Config\LogLevel;
 use mteu\StreamWriter\Config\StandardStream;
 use mteu\StreamWriter\Exception\InvalidLogWriterConfigurationException;
 use mteu\StreamWriter\Exception\InvalidLogWriterOptionException;
-use mteu\StreamWriter\ValueObjects\Message;
-use mteu\StreamWriter\ValueObjects\MessageFactory;
+use mteu\StreamWriter\Log\Message\Message;
+use mteu\StreamWriter\Log\Message\MessageFactory;
 use TYPO3\CMS\Core\Log\LogRecord;
 use TYPO3\CMS\Core\Log\Writer\AbstractWriter;
 use TYPO3\CMS\Core\Log\Writer\WriterInterface;
